Revert "Added .gitignore"
This reverts commit aa80b534eebac1e5082770e25d36ae1632e703ba.
