--- .gitignore
@@ -1,3 +1,0 @@
-repo_mining/actual key.txt
-repo_mining/data/CollectFiles_file_rootbeer.csv
-repo_mining/data/file_rootbeer.csv
